pyproject.toml
[project]
name = "im-automated-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "playwright>=1.51.0",
    "pre-commit>=4.2.0",
    "pytest-base-url>=2.1.0",
    "pytest-flakefinder>=1.1.0",
    "pytest-playwright>=0.7.0",
    "pytest-xdist>=3.6.1",
    "python-dotenv>=1.1.0",
]


[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]
